реалізація транзакцій між рахунками різних клієнтів та обчислення загального балансу Банка
