+++ README.md
@@ -90,7 +90,7 @@ accounts (список всіх рахунків банку).
 
  "-" Проведення транзакцій між рахунками різних клієнтів.
 
- "-" Обчислення загальних активів банку.
+ "+" Обчислення загальних активів банку.
 
  "-" Генерація виписок по рахунках клієнтів.
