+++ bank.py
@@ -30,14 +30,23 @@ class Bank:
         client_account = (ClientFactory().
                           create_account(account_type=account_type,
                                          client=client))
-        self.add_account(client_account)
+        self.add_account_to_list(client_account)
         return client_account
+
+    @staticmethod
+    def transfer_between_client_account(with_account, to_account, amount):
+        """Проведення транзакцій між рахунками різних клієнтів"""
+        with_account.transfer(to_account, amount)
 
     def add_client_to_list(self, client):
         self.clients.append(client)
 
-    def add_account(self, account):
+    def add_account_to_list(self, account):
         self.accounts.append(account)
+
+    def get_total_balance(self):
+        """Обчислення загальних активів банку"""
+        return sum(account.balance for account in self.accounts)
 
     def get_list_clients(self):
         for index, client in enumerate(self.clients):
