+++ main.py
@@ -1,10 +1,9 @@
 from bank import Bank
 from client import Client, ClientFactory
-from account import AccountFactory
 from faker import Faker
 
 
-fake = Faker()
+fake = Faker('uk_UA')
 
 my_bank = Bank()
 print('---Створення нового клієнта---\n')
@@ -17,9 +16,21 @@ print('\n---Збільшення кредитного ліміту---\n')
 credit_account.set_credit_limit(100000)
 print(client)
 print('\n---Поповнення ощадного рахунку з кредитного---\n')
-account_savings = client.list_accounts[0]
 credit_account.transfer(account_savings, 60000)
 print(credit_account, account_savings)
-# my_bank.get_list_clients()
-# my_bank.get_list_accounts()
+print('\n---Нарахування відсотків на кредитний рахунок---\n')
+credit_account.apply_loan_interest()
+print('\n---Створення ще одного клієнта ---\n')
+client_two = my_bank.create_new_client(fake.name())
+account_savings_two = my_bank.open_new_account_client(client_two, 'savings')
+account_savings_two.increase_balance(100000)
+print(client_two)
+print('\n---Переведення коштів від другого клієнта першому ---\n')
+my_bank.transfer_between_client_account(account_savings_two, credit_account, 60000)
 
+print('First client\n', credit_account)
+print('Second client\n', account_savings_two)
+
+print('\n---Загальний баланс банку ---\n')
+my_bank.get_list_accounts()
+print(my_bank.get_total_balance(), '$')
